+++ api/server/user/internal/domain/user/repository.go
@@ -13,15 +13,15 @@ type Repository interface {
 	ListFollow(ctx context.Context, q *domain.ListQuery) ([]*Follow, error)
 	ListFollower(ctx context.Context, q *domain.ListQuery) ([]*Follower, error)
 	ListFollowerID(ctx context.Context, q *domain.ListQuery) ([]string, error)
-	ListCount(ctx context.Context, q *domain.ListQuery) (int64, error)
-	ListRelationshipCount(ctx context.Context, q *domain.ListQuery) (int64, error)
+	ListCount(ctx context.Context, q *domain.ListQuery) (int, error)
+	ListRelationshipCount(ctx context.Context, q *domain.ListQuery) (int, error)
 	Show(ctx context.Context, uid string) (*User, error)
-	ShowRelationship(ctx context.Context, id int64) (*Relationship, error)
+	ShowRelationship(ctx context.Context, id int) (*Relationship, error)
 	Create(ctx context.Context, u *User) error
 	CreateRelationship(ctx context.Context, r *Relationship) error
 	Update(ctx context.Context, u *User) error
 	UpdatePassword(ctx context.Context, uid string, password string) error
-	DeleteRelationship(ctx context.Context, id int64) error
+	DeleteRelationship(ctx context.Context, id int) error
 	GetUIDByEmail(ctx context.Context, email string) (string, error)
-	GetRelationshipIDByUID(ctx context.Context, followID string, followerID string) (int64, error)
+	GetRelationshipIDByUID(ctx context.Context, followID string, followerID string) (int, error)
 }
